\name{QCSIS-package}
\alias{QCSIS-package}
\docType{package}
\title{
\packageTitle{QCSIS}
}
\description{
\packageDescription{QCSIS}
}
\details{

\packageDESCRIPTION{QCSIS}
\packageIndices{QCSIS}

}
\author{
\packageAuthor{QCSIS}

Maintainer: \packageMaintainer{QCSIS}
}
\references{
Xuejun Ma  and Jingxiao Zhang. Robust model-free feature screening via quantile correlation. Journal of Multivariate Analysis. Online, 2015.\cr
 Xuejun Ma et al.. Robust feature screening via composite quantile correlation learning.   In submission. \cr

}

